requirements/base.txt
diskcache==4.1.0
django-js-asset<2
django-filter==1.1.0 # pyup: <2.0.0
django-js-reverse==0.9.1
djangorestframework==3.9.1
django==1.11.29  # pyup: >=1.11,<2
six==1.14.0
colorlog==3.2.0 # pyup: <4.0.0
configobj==5.0.6
django-mptt==0.9.1
requests==2.21.0
cheroot==8.5.2
magicbus==4.1.2
futures==3.1.1  # Temporarily pinning this until we can do a Python 2/3 compatible solution of newer versions # pyup: <=3.1.1
more-itertools==5.0.0  # Last Python 2.7 friendly release # pyup: <6.0
le-utils==0.1.38
jsonfield==2.0.2
requests-toolbelt==0.8.0
morango==0.6.10
tzlocal==2.1
pytz==2020.5
python-dateutil==2.7.5
sqlalchemy==1.3.17
semver==2.8.1
django-redis-cache==2.0.0
redis==3.2.1
html5lib==1.0.1
zeroconf-py2compat==0.19.16
ifcfg==0.21
Click==7.0
whitenoise==4.1.4
idna==2.8
